requirements.txt
pandas
huggingface_hub
notion-client
python-dotenv